Move new forms to next to their uses
