+++ packages/desktop/app/forms/index.js
@@ -1,9 +1,7 @@
 export { AllergyForm } from './AllergyForm';
-export { EditPatientLetterTemplateForm } from './EditPatientLetterTemplateForm';
 export { FamilyHistoryForm } from './FamilyHistoryForm';
 export { NewLocationForm } from './NewLocationForm';
 export { NewPatientForm } from './NewPatientForm';
-export { NewTemplateForm } from './NewTemplateForm';
 export { NewUserForm } from './NewUserForm';
 export { OngoingConditionForm } from './OngoingConditionForm';
 export { PatientIssueForm } from './PatientIssueForm';
